Add handler for retrieving trusted source signup details
