+++ .coveragerc
@@ -1,0 +1,2 @@
+[run]
+omit = .venv/*,setup.py
