+++ directory_api_client/dummy_client.py
@@ -10,7 +10,7 @@ class DummyDirectoryAPIClient(DirectoryAPIClient):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        patch.object(self.registration, 'send', self.send).start()
+        patch.object(self.enrolment, 'send', self.send).start()
         patch.object(self.supplier, 'send', self.send).start()
         patch.object(self.company, 'send', self.send).start()
         patch.object(self.buyer, 'send', self.send).start()
