+++ directory_api_client/version.py
@@ -1,1 +1,1 @@
-__version__ = '4.5.0'
+__version__ = '5.0.0'
